move submit_to_loop to tasks and remove_schedule
